+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
-import { BudgetProvider } from './context/BudgetContext.tsx';
+import { BudgetProvider } from './context/BudgetProvider.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
